Simple_RAG/requirements.txt
numpy>=1.24
requests>=2.32
pypdf>=4.2
fastapi>=0.112
uvicorn>=0.30
python-dotenv>=1.0
python-multipart>=0.0.20
google-cloud-aiplatform>=1.66.0
google-auth>=2.35.0
google-cloud-storage>=2.18.0
google-api-core>=2.19.2